Fix problem where folder would not open if one replaces an existing file (#46)
If one choose to use saveAs behaviour and the user downloads a file for the second time, the code for opening that file after the download has completed does not work. This fix will always use name of the actual file that was downloaded, not some initial variable set up front.