Fixing The Progressbar Problem
The problem occurs when you try to download multiple files, and the progress gets corrupted, The code prevents re-emitting the callbacks with the same options with multiple progress. 

The issue is resolved by @faustor21, I am just making a pull request so that this can be implimneted. All the props to him.